+++ src/Symfony/Bundle/FrameworkBundle/Tests/Kernel/MicroKernelTraitTest.php
@@ -28,4 +28,15 @@ class MicroKernelTraitTest extends TestCase
         $this->assertEquals('Have a great day!', $kernel->getContainer()->getParameter('halloween'));
         $this->assertInstanceOf('stdClass', $kernel->getContainer()->get('halloween'));
     }
+
+    public function testAsEventSubscriber()
+    {
+        $kernel = new ConcreteMicroKernel('test', true);
+        $kernel->boot();
+
+        $request = Request::create('/danger');
+        $response = $kernel->handle($request);
+
+        $this->assertSame('It\'s dangerous to go alone. Take this ⚔', $response->getContent());
+    }
 }