+++ src/Symfony/Component/HttpKernel/CHANGELOG.md
@@ -24,6 +24,7 @@ CHANGELOG
 3.4.0
 -----
 
+ * deprecated commands auto registration
  * added `AddCacheClearerPass`
  * added `AddCacheWarmerPass`